rename resize helpers, export getLockedAspectRatio
